fix for indexRandom recalculation, now only when return to 0

+++ src/Nodes/Default_Nodes/Base/baseIndexer.h
@@ -52,7 +52,7 @@ private:
     
     
     vector<int>         indexRand;
-    int                 indexRand_Param_previous;
+    float               indexRand_Param_previous;
     bool newIndexsFlag;
     int previousIndexCount;
 };
